printf

+++ di.c
@@ -6,7 +6,7 @@
 /*   By: jinfeld <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/08/23 17:06:10 by jinfeld           #+#    #+#             */
-/*   Updated: 2017/08/29 08:35:23 by jinfeld          ###   ########.fr       */
+/*   Updated: 2017/09/12 16:19:08 by jinfeld          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ char	*di(va_list args, char c, int mod)
 	if (mod == 1 || c == 'D')
 		str = ft_itoa(va_arg(args, long));
 	if (mod == 2)
-		str = ft_itoa((intmax_t)va_arg(args, long long int));
+		str = ft_itoa(va_arg(args, long long int));
 	if (mod == 3)
 		str = ft_itoa((short)va_arg(args, int));
 	if (mod == 4)
